Fix stupid default permanent storage path
As pointed out by @rvetere, `/tmp/owncloud` is one of the worst default
paths for the persistent data of the container created using the
Makefile. What makes this worse is that the documentation did not
mention it.

This default value was chosen in the assumption that everybody would
review at least the commands that are executed by make that would show
that all volumes are bind mounted from `/tmp/`. Of course this
assumption is bad so thanks very much to @rvetere for saying something!

Related to: https://github.com/jchaney/owncloud/issues/58

I will leave this issue open so that people who might have such a setup
running hopefully notice the issue before it is too late.